LayoutTests/editing/pasteboard/entries-api/DirectoryReader-readEntries-back-forward-cache-expected.txt

Basic test coverage for fileSystemDirectoryReader.readEntries()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


pageshow - not from cache
pagehide - entering cache
pageshow - from cache
PASS Page did enter and was restored from the page cache
PASS Success callback was called
PASS restoredFromPageCache is true
PASS successfullyParsed is true

TEST COMPLETE
